Change how we handle OpenShift strategies
When selecting an OpenShift strategy, the implementation handler for the managed resource will be selected depending on the chosen OpenShift strategy.

Previously, this selection was done upon dynamically using Service Loaders.
With these changes, we won't allow to add new OpenShift strategies via the extension model.
Still users will be allowed to extend the framework via adding a new annotation.